# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: A-Level Revision Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List the shared public subject catalogue
      description: |
        Returns importer/admin-managed reference subjects. Not user-scoped.
        Personal-looking fields on this legacy shared response are intentionally neutral
        (syllabusProgress/topicsCompleted/topicsInProgress = 0; upcomingTasksCount = 0;
        recent paper fields null). Caller-owned data is exposed by authenticated endpoints.
      responses:
        "200":
          description: Shared catalogue subjects with neutral placeholder enrichment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Disabled — subject catalogue is read-only shared reference data
      description: |
        Ordinary users must not insert into the global catalogue.
        Subjects are populated by the syllabus importer / admin processes.
        No admin role is exposed in this API version. No database write is performed.
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "403":
          description: Catalogue writes are forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get a shared catalogue subject with neutral placeholder enrichment
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject detail (progress/task/paper fields are placeholders)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Not found
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Disabled — subject catalogue is read-only shared reference data
      description: |
        Ordinary users must not delete catalogue rows. No database delete is performed.
      deprecated: true
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "403":
          description: Catalogue deletes are forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
  /subjects/{subjectId}/syllabus:
    get:
      operationId: getSubjectSyllabus
      tags: [subjects]
      summary: Get syllabus reference structure for a subject
      description: |
        Returns units, topic titles, and learning outcomes from shared reference data.
        When a valid Bearer token is provided, each topic's status and notes are merged
        from the caller's topic_progress rows. Missing progress rows default to
        not_started with null notes. Unauthenticated callers receive the same defaults.
        Progress fields come from topic_progress, not shared catalogue columns.
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Syllabus units with topics (caller progress when authenticated)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SyllabusUnit"
        "401":
          description: Authorization header present but invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /subjects/{subjectId}/performance:
    get:
      operationId: getSubjectPerformance
      tags: [subjects]
      summary: Get caller-owned past-paper performance for one subject
      description: |
        Requires Bearer auth. Latest, average, best, count, trend and component
        breakdown are calculated only from the caller's attempts.
      security:
        - bearerAuth: []
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Caller-owned subject performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectPerformance"
        "404":
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /subjects/{subjectId}/syllabus-versions/{syllabusVersionId}/assessment-routes:
    get:
      operationId: listSubjectAssessmentRoutes
      tags: [subjects]
      summary: List published assessment routes for a syllabus version
      description: |
        Returns the published route set for the given syllabus version of a subject,
        including routes, components, and study-option groups. Fails closed when a
        published route set is required but missing or ambiguous. When no published
        route set exists yet, returns an empty routes array (legacy-compatible).
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: syllabusVersionId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Published route catalogue for the version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectAssessmentRouteCatalogue"
        "400":
          description: Invalid parameters or ambiguous route set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: Subject or syllabus version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /subjects/{subjectId}/assessment-components:
    get:
      operationId: listAssessmentComponents
      tags: [subjects]
      summary: List assessment components (papers) for a subject's current syllabus version, for the Log Paper Component dropdown
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of assessment components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentComponent"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token (Authorization Bearer) for authenticated caller-owned routes.
  schemas:
    ErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SubjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        color:
          type: string
          minLength: 1
      required:
        - name
        - code
        - color
    Subject:
      type: object
      description: |
        Shared catalogue subject. User-specific progress, task-count, and past-paper
        fields remain neutral placeholders on this public catalogue response.
        Catalogue list endpoints return only subjects selectable for new memberships.
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        color:
          type: string
        selectableForNewMemberships:
          type: boolean
          description: |
            When false, omitted from new-membership catalogue surfaces. Existing
            memberships remain accessible via membership APIs.
        syllabusProgress:
          type: number
          description: Neutral placeholder; always 0 (not derived from shared topic status)
        topicsTotal:
          type: integer
          description: Count of syllabus topics in the shared catalogue for this subject
        topicsCompleted:
          type: integer
          description: Neutral placeholder; always 0
        topicsInProgress:
          type: integer
          description: Neutral placeholder; always 0
        upcomingTasksCount:
          type: integer
          description: Neutral placeholder; always 0 (not user-task derived)
        recentPaperScore:
          type: ["number", "null"]
          description: Neutral placeholder; always null on shared catalogue responses
        recentPaperLabel:
          type: ["string", "null"]
          description: Neutral placeholder; always null on shared catalogue responses
      required:
        - id
        - name
        - code
        - color
        - selectableForNewMemberships
        - syllabusProgress
        - topicsTotal
        - topicsCompleted
        - topicsInProgress
        - upcomingTasksCount
        - recentPaperScore
        - recentPaperLabel
    SyllabusUnit:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        title:
          type: string
        orderIndex:
          type: integer
        topics:
          type: array
          items:
            $ref: "#/components/schemas/SyllabusTopic"
      required:
        - id
        - subjectId
        - title
        - orderIndex
        - topics
    SyllabusTopic:
      type: object
      description: |
        Shared syllabus topic reference with caller-owned progress fields merged
        when authenticated. Missing topic_progress rows default to not_started
        and null notes. Progress fields come from topic_progress only.
      properties:
        id:
          type: integer
        unitId:
          type: integer
        subjectId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed]
          description: Caller-owned progress status; defaults to not_started
        notes:
          type: ["string", "null"]
          description: Caller-owned notes; defaults to null
        orderIndex:
          type: integer
        learningOutcomes:
          type: array
          items:
            type: string
      required:
        - id
        - unitId
        - subjectId
        - title
        - status
        - notes
        - orderIndex
        - learningOutcomes
...
    AssessmentComponent:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        paperCode:
          type: string
        level:
          type: string
        componentName:
          type: string
        durationMinutes:
          type: ["integer", "null"]
        totalMarks:
          type: ["integer", "null"]
        weightingPercent:
          type: ["number", "null"]
        orderIndex:
          type: integer
      required:
        - id
        - subjectId
        - paperCode
        - level
        - componentName
        - durationMinutes
        - totalMarks
        - weightingPercent
        - orderIndex
...
    SubjectPerformance:
      type: object
      properties:
        subjectId:
          type: integer
        subjectName:
          type: string
        latestScore:
          type: ["number", "null"]
        averageScore:
          type: ["number", "null"]
        bestScore:
          type: ["number", "null"]
        papersCompleted:
          type: integer
        trend:
          type: array
          items:
            $ref: "#/components/schemas/PaperTrendPoint"
        componentBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ComponentBreakdown"
        insight:
          type: ["string", "null"]
      required:
        - subjectId
        - subjectName
        - latestScore
        - averageScore
        - bestScore
        - papersCompleted
        - trend
        - componentBreakdown
        - insight
    PaperTrendPoint:
      type: object
      properties:
        label:
          type: string
        percentage:
          type: number
        session:
          type: string
      required:
        - label
        - percentage
        - session
    ComponentBreakdown:
      type: object
      properties:
        componentId:
          type: ["integer", "null"]
        componentName:
          type: string
        latestPercentage:
          type: ["number", "null"]
        attempts:
          type: integer
      required:
        - componentId
        - componentName
        - latestPercentage
        - attempts
...
    SubjectAssessmentRouteCatalogue:
      type: object
      properties:
        subjectId:
          type: integer
        syllabusVersionId:
          type: integer
        routeSetId:
          type: ["integer", "null"]
        routeRevisionKey:
          type: ["string", "null"]
        selectionMode:
          type: string
          enum: [none_available, auto, explicit]
        routes:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentRouteSummary"
        optionGroups:
          type: array
          items:
            $ref: "#/components/schemas/StudyOptionGroupSummary"
      required:
        - subjectId
        - syllabusVersionId
        - routeSetId
        - routeRevisionKey
        - selectionMode
        - routes
        - optionGroups
    AssessmentRouteSummary:
      type: object
      properties:
        id:
          type: integer
        routeKey:
          type: string
        displayLabel:
          type: string
        qualificationTarget:
          type: string
          enum: [as_level, a_level]
        pathwayType:
          type: string
          enum: [single_series, staged_completion, full_same_series]
        progressionEligibility:
          type: string
          enum: [eligible, not_eligible, not_applicable]
        orderIndex:
          type: integer
        components:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentRouteComponentSummary"
      required:
        - id
        - routeKey
        - displayLabel
        - qualificationTarget
        - pathwayType
        - progressionEligibility
        - orderIndex
        - components
    AssessmentRouteComponentSummary:
      type: object
      properties:
        componentId:
          type: integer
        paperCode:
          type: string
        componentLabel:
          type: string
        role:
          type: string
          enum: [current_sitting, carried_forward]
        orderIndex:
          type: integer
      required:
        - componentId
        - paperCode
        - componentLabel
        - role
        - orderIndex
    StudyOptionGroupSummary:
      type: object
      properties:
        id:
          type: integer
        groupKey:
          type: string
        displayLabel:
          type: string
        minSelections:
          type: integer
        maxSelections:
          type: integer
        orderIndex:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/StudyOptionSummary"
      required:
        - id
        - groupKey
        - displayLabel
        - minSelections
        - maxSelections
        - orderIndex
        - options
    StudyOptionSummary:
      type: object
      properties:
        id:
          type: integer
        optionKey:
          type: string
        displayLabel:
          type: string
        orderIndex:
          type: integer
      required:
        - id
        - optionKey
        - displayLabel
        - orderIndex