# lib/api-zod/src/generated/types/assessmentRouteComponentSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * A-Level Revision Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentRouteComponentSummaryRole } from './assessmentRouteComponentSummaryRole';

export interface AssessmentRouteComponentSummary {
  componentId: number;
  paperCode: string;
  componentLabel: string;
  role: AssessmentRouteComponentSummaryRole;
  orderIndex: number;
}