# lib/api-zod/src/generated/types/assessmentRouteSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * A-Level Revision Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentRouteComponentSummary } from './assessmentRouteComponentSummary';
import type { AssessmentRouteSummaryPathwayType } from './assessmentRouteSummaryPathwayType';
import type { AssessmentRouteSummaryProgressionEligibility } from './assessmentRouteSummaryProgressionEligibility';
import type { AssessmentRouteSummaryQualificationTarget } from './assessmentRouteSummaryQualificationTarget';

export interface AssessmentRouteSummary {
  id: number;
  routeKey: string;
  displayLabel: string;
  qualificationTarget: AssessmentRouteSummaryQualificationTarget;
  pathwayType: AssessmentRouteSummaryPathwayType;
  progressionEligibility: AssessmentRouteSummaryProgressionEligibility;
  orderIndex: number;
  components: AssessmentRouteComponentSummary[];
}