# lib/api-zod/src/generated/types/studyOptionGroupSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * A-Level Revision Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StudyOptionSummary } from './studyOptionSummary';

export interface StudyOptionGroupSummary {
  id: number;
  groupKey: string;
  displayLabel: string;
  minSelections: number;
  maxSelections: number;
  orderIndex: number;
  options: StudyOptionSummary[];
}