# lib/api-zod/src/generated/types/subjectAssessmentRouteCatalogue.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * A-Level Revision Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentRouteSummary } from './assessmentRouteSummary';
import type { StudyOptionGroupSummary } from './studyOptionGroupSummary';
import type { SubjectAssessmentRouteCatalogueSelectionMode } from './subjectAssessmentRouteCatalogueSelectionMode';

export interface SubjectAssessmentRouteCatalogue {
  subjectId: number;
  syllabusVersionId: number;
  /** @nullable */
  routeSetId: number | null;
  /** @nullable */
  routeRevisionKey: string | null;
  selectionMode: SubjectAssessmentRouteCatalogueSelectionMode;
  routes: AssessmentRouteSummary[];
  optionGroups: StudyOptionGroupSummary[];
}